extract nested TestingMetricGroup as a stand-alone class since it is big enough.
